Update README.md
A/B test

The purpose of this project is to understand the results of an A/B test run by an e-commerce website. 
The company has developed a new web page in order to try and increase the number of users who "convert," meaning the number of users who decide to pay for the company's product.
The goal is to work through the A/B test analaysis to help the company understand if they should implement this new page, keep the old page, or perhaps run the experiment longer to make their decision.